Refactor explosion data into modules
