+++ src/db/explosions-db.ts
@@ -1,1095 +1,1 @@
-import {
-  SceneColor,
-  SceneFill,
-  SceneFillFilaments,
-  SceneFillNoise,
-  SceneGradientStop,
-  SceneVector2,
-} from "@core/logic/provided/services/scene-object-manager/scene-object-manager.types";
-import { FILL_TYPES } from "@core/logic/provided/services/scene-object-manager/scene-object-manager.const";
-import { getAssetUrl } from "@shared/helpers/assets.helper";
-import type { ParticleEmitterConfig } from "../logic/interfaces/visuals/particle-emitters-config";
-
-export type ExplosionType =
-  | "plasmoid"
-  | "magnetic"
-  | "healWave"
-  | "fireball"
-  | "grayBrickHit"
-  | "grayBrickDestroy"
-  | "grayBrickDestroyV2"
-  | "yellowBrickHit"
-  | "yellowBrickDestroy"
-  | "organicBrickHit"
-  | "organicBrickDestroy"
-  | "ironBrickHit"
-  | "ironBrickDestroy"
-  | "woodBrickHit"
-  | "woodBrickDestroy"
-  | "copperBrickHit"
-  | "copperBrickDestroy"
-  | "silverBrickHit"
-  | "silverBrickDestroy"
-  | "coalBrickHit"
-  | "coalBrickDestroy"
-  | "iceBrickHit"
-  | "iceBrickDestroy"
-  | "magmaBrickHit"
-  | "magmaBrickDestroy"
-  | "criticalHit"
-  | "weakenCurse"
-  | "smallCannon"
-  | "bigCannon"
-  | "smallCannonGrey"
-  | "smallLaser"
-  | "smallEnergetic";
-
-export interface ExplosionWaveConfig {
-  initialInnerRadius: number;
-  expansionInnerRadius: number;
-  initialOuterRadius: number;
-  expansionOuterRadius: number;
-  startAlpha: number;
-  endAlpha: number;
-  gradientStops: readonly SceneGradientStop[];
-  noise?: SceneFillNoise;
-  filaments?: SceneFillFilaments;
-}
-
-export interface ExplosionConfig {
-  lifetimeMs: number;
-  defaultInitialRadius: number;
-  waves: readonly ExplosionWaveConfig[];
-  emitter: ParticleEmitterConfig;
-  soundEffectUrl?: string;
-}
-
-const createSimpleWave = (options: {
-  defaultInitialRadius: number;
-  radiusExtension: number;
-  startAlpha: number;
-  endAlpha: number;
-  gradientStops: readonly SceneGradientStop[];
-}): ExplosionWaveConfig[] => [
-  {
-    initialInnerRadius: 0,
-    expansionInnerRadius: 0,
-    initialOuterRadius: options.defaultInitialRadius,
-    expansionOuterRadius:
-      options.defaultInitialRadius + Math.max(0, options.radiusExtension),
-    startAlpha: options.startAlpha,
-    endAlpha: options.endAlpha,
-    gradientStops: options.gradientStops,
-  },
-];
-
-const PLASMOID_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 0.75, b: 0.3, a: 0.8 } },
-  { offset: 0.35, color: { r: 1, g: 0.45, b: 0.15, a: 0.55 } },
-  { offset: 1, color: { r: 1, g: 0.1, b: 0, a: 0 } },
-] as const;
-
-const GREY_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.6, g: 0.75, b: 0.75, a: 0.8 } },
-  { offset: 0.35, color: { r: 0.6, g: 0.75, b: 0.75, a: 0.55 } },
-  { offset: 1, color: { r: 0.6, g: 0.75, b: 0.75, a: 0 } },
-] as const;
-
-const MAGNETIC_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.6, g: 0.4, b: 1, a: 0.85 } },
-  { offset: 0.45, color: { r: 0.45, g: 0.2, b: 0.95, a: 0.6 } },
-  { offset: 1, color: { r: 0.25, g: 0.05, b: 0.7, a: 0 } },
-] as const;
-
-const GRAY_BRICK_HIT_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.92, g: 0.92, b: 0.94, a: 0.3 } },
-  { offset: 0.4, color: { r: 0.75, g: 0.77, b: 0.8, a: 0.6 } },
-  { offset: 1, color: { r: 0.55, g: 0.58, b: 0.6, a: 0 } },
-] as const;
-
-const GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.95, g: 0.95, b: 0.97, a: 0.5 } },
-  { offset: 0.35, color: { r: 0.78, g: 0.8, b: 0.84, a: 0.8 } },
-  { offset: 1, color: { r: 0.45, g: 0.48, b: 0.52, a: 0 } },
-] as const;
-
-const CRITICAL_HIT_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 0.35, b: 0.35, a: 0.5 } },
-  { offset: 0.4, color: { r: 0.9, g: 0.12, b: 0.12, a: 0.8 } },
-  { offset: 1, color: { r: 0.55, g: 0, b: 0, a: 0 } },
-] as const;
-
-const HEAL_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.55, g: 1.0, b: 0.65, a: 0.5 } },
-  { offset: 0.4, color: { r: 0.35, g: 0.9, b: 0.45, a: 0.8 } },
-  { offset: 1, color: { r: 0.1, g: 0.6, b: 0.15, a: 0 } },
-] as const;
-
-// Color-themed waves for non-gray bricks
-const YELLOW_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.95, g: 0.92, b: 0.55, a: 0.3 } },
-  { offset: 0.4, color: { r: 0.85, g: 0.8, b: 0.35, a: 0.5 } },
-  { offset: 1, color: { r: 0.65, g: 0.6, b: 0.2, a: 0 } },
-] as const;
-const YELLOW_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 0.95, b: 0.6, a: 0.5 } },
-  { offset: 0.35, color: { r: 0.9, g: 0.82, b: 0.4, a: 0.8 } },
-  { offset: 1, color: { r: 0.7, g: 0.6, b: 0.25, a: 0 } },
-] as const;
-
-const GREEN_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.6, g: 0.95, b: 0.65, a: 0.15 } },
-  { offset: 0.4, color: { r: 0.35, g: 0.85, b: 0.45, a: 0.4 } },
-  { offset: 1, color: { r: 0.1, g: 0.55, b: 0.15, a: 0 } },
-] as const;
-const GREEN_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.6, g: 1, b: 0.7, a: 0.3 } },
-  { offset: 0.35, color: { r: 0.4, g: 0.9, b: 0.5, a: 0.5 } },
-  { offset: 1, color: { r: 0.1, g: 0.6, b: 0.15, a: 0 } },
-] as const;
-
-const ORANGE_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 0.8, b: 0.55, a: 0.3 } },
-  { offset: 0.4, color: { r: 0.95, g: 0.55, b: 0.2, a: 0.5 } },
-  { offset: 1, color: { r: 0.75, g: 0.35, b: 0.1, a: 0 } },
-] as const;
-const ORANGE_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 0.85, b: 0.6, a: 0.5 } },
-  { offset: 0.35, color: { r: 0.95, g: 0.6, b: 0.25, a: 0.8 } },
-  { offset: 1, color: { r: 0.8, g: 0.4, b: 0.12, a: 0 } },
-] as const;
-
-const BROWN_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.7, g: 0.5, b: 0.3, a: 0.3 } },
-  { offset: 0.4, color: { r: 0.55, g: 0.4, b: 0.2, a: 0.5 } },
-  { offset: 1, color: { r: 0.35, g: 0.25, b: 0.12, a: 0 } },
-] as const;
-const BROWN_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.75, g: 0.55, b: 0.35, a: 0.5 } },
-  { offset: 0.35, color: { r: 0.6, g: 0.45, b: 0.25, a: 0.8 } },
-  { offset: 1, color: { r: 0.4, g: 0.3, b: 0.15, a: 0 } },
-] as const;
-
-const SILVER_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.92, g: 0.93, b: 0.96, a: 0.3 } },
-  { offset: 0.45, color: { r: 0.82, g: 0.83, b: 0.88, a: 0.5 } },
-  { offset: 1, color: { r: 0.72, g: 0.73, b: 0.78, a: 0 } },
-] as const;
-const SILVER_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1, g: 1, b: 1, a: 0.5 } },
-  { offset: 0.45, color: { r: 0.85, g: 0.87, b: 0.9, a: 0.8 } },
-  { offset: 1, color: { r: 0.75, g: 0.76, b: 0.8, a: 0 } },
-] as const;
-
-const COAL_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.3, g: 0.3, b: 0.32, a: 0.3 } },
-  { offset: 0.5, color: { r: 0.18, g: 0.18, b: 0.2, a: 0.5 } },
-  { offset: 1, color: { r: 0.08, g: 0.08, b: 0.1, a: 0 } },
-] as const;
-const COAL_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.36, g: 0.36, b: 0.38, a: 0.5 } },
-  { offset: 0.5, color: { r: 0.22, g: 0.22, b: 0.25, a: 0.8 } },
-  { offset: 1, color: { r: 0.1, g: 0.1, b: 0.12, a: 0 } },
-] as const;
-
-// Ice-themed waves (cool cyan/white)
-const ICE_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.7, g: 0.9, b: 1.0, a: 0.5 } },
-  { offset: 0.45, color: { r: 0.55, g: 0.8, b: 1.0, a: 0.3 } },
-  { offset: 1, color: { r: 0.35, g: 0.6, b: 0.9, a: 0 } },
-] as const;
-const ICE_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.85, g: 0.95, b: 1.0, a: 0.65 } },
-  { offset: 0.45, color: { r: 0.65, g: 0.85, b: 1.0, a: 0.4 } },
-  { offset: 1, color: { r: 0.4, g: 0.7, b: 1.0, a: 0 } },
-] as const;
-
-// Magma-themed waves (hot orange/red)
-const MAGMA_BRICK_HIT_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1.0, g: 0.5, b: 0.15, a: 0.5 } },
-  { offset: 0.45, color: { r: 0.9, g: 0.3, b: 0.1, a: 0.28 } },
-  { offset: 1, color: { r: 0.7, g: 0.15, b: 0.05, a: 0 } },
-] as const;
-const MAGMA_BRICK_DESTROY_WAVE: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 1.0, g: 0.65, b: 0.2, a: 0.68 } },
-  { offset: 0.45, color: { r: 1.0, g: 0.45, b: 0.15, a: 0.4 } },
-  { offset: 1, color: { r: 0.85, g: 0.25, b: 0.08, a: 0 } },
-] as const;
-
-const DEFAULT_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.SOLID,
-  color: { r: 1, g: 0.85, b: 0.55, a: 1 },
-};
-
-
-const SMALL_CANNON_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 1.0, g: 0.9, b: 0.8, a: 1.0 } },
-    { offset: 0.45, color: { r: 0.9, g: 0.8, b: 0.7, a: 0.42 } },
-    { offset: 1, color: { r: 0.5, g: 0.4, b: 0.3, a: 0.02 } },
-  ]
-};
-
-const SMALL_LASER_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 1.0, g: 0.7, b: 0.8, a: 1.0 } },
-    { offset: 0.45, color: { r: 0.9, g: 0.6, b: 0.7, a: 0.42 } },
-    { offset: 1, color: { r: 0.9, g: 0.5, b: 0.6, a: 0.32 } },
-  ]
-};
-
-const SMALL_ENERGETIC_WAVE_GRADIENT_STOPS: readonly SceneGradientStop[] = [
-  { offset: 0, color: { r: 0.5, g: 0.8, b: 0.8, a: 0.4 } },
-  { offset: 0.75, color: { r: 0.4, g: 0.7, b: 0.6, a: 0.9 } },
-  { offset: 1, color: { r: 0.4, g: 0.7, b: 0.6, a: 0 } },
-] as const;
-
-const SMALL_ENERGETIC_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.7, g: 1.0, b: 1.0, a: 1.0 } },
-    { offset: 0.45, color: { r: 0.4, g: 0.9, b: 0.8, a: 0.6 } },
-    { offset: 1, color: { r: 0.4, g: 0.9, b: 0.8, a: 0.2 } },
-  ]
-};
-
-const SMALL_GREY_CANNON_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.6, g: 0.75, b: 0.75, a: 0.8 } },
-    { offset: 0.35, color: { r: 0.6, g: 0.75, b: 0.75, a: 0.55 } },
-    { offset: 1, color: { r: 0.6, g: 0.75, b: 0.75, a: 0 } },
-  ],
-};
-
-
-const MAGNETIC_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  end: 8,
-  stops: [
-    { offset: 0, color: { r: 1, g: 1, b: 1, a: 0.25 } },
-    { offset: 0.4, color: { r: 0.8, g: 0.6, b: 1, a: 0.2 } },
-    { offset: 1, color: { r: 0.6, g: 0.4, b: 0.9, a: 0.05 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.05,
-    scale: 0.3,
-  },
-};
-
-const GRAY_BRICK_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.9, g: 0.9, b: 0.92, a: 0.12 } },
-    { offset: 0.45, color: { r: 0.72, g: 0.74, b: 0.78, a: 0.06 } },
-    { offset: 1, color: { r: 0.45, g: 0.48, b: 0.52, a: 0.03 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.075,
-    scale: 0.5,
-    density: 0.3,
-  },
-};
-const YELLOW_BRICK_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 1, g: 0.95, b: 0.6, a: 0.09 } },
-    { offset: 0.45, color: { r: 0.95, g: 0.85, b: 0.45, a: 0.05 } },
-    { offset: 1, color: { r: 0.8, g: 0.7, b: 0.35, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.02,
-    alphaAmplitude: 0.06,
-    scale: 0.5,
-    density: 0.3,
-  },
-};
-const GREEN_BRICK_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.4, g: 0.8, b: 0.5, a: 0.12 } },
-    { offset: 0.45, color: { r: 0.2, g: 0.6, b: 0.3, a: 0.06 } },
-    { offset: 1, color: { r: 0.05, g: 0.45, b: 0.1, a: 0.03 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-    density: 0.2,
-  },
-};
-const ORANGE_BRICK_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 1, g: 0.85, b: 0.6, a: 0.12 } },
-    { offset: 0.45, color: { r: 1, g: 0.6, b: 0.25, a: 0.06 } },
-    { offset: 1, color: { r: 0.85, g: 0.4, b: 0.12, a: 0.03 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-    density: 0.2,
-  },
-};
-const BROWN_BRICK_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.75, g: 0.55, b: 0.35, a: 0.09 } },
-    { offset: 0.45, color: { r: 0.6, g: 0.45, b: 0.25, a: 0.05 } },
-    { offset: 1, color: { r: 0.4, g: 0.3, b: 0.15, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-    density: 0.3,
-  },
-};
-const SILVER_BRICK_EMITTER_FILL2: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.95, g: 0.96, b: 0.98, a: 0.09 } },
-    { offset: 0.45, color: { r: 0.85, g: 0.86, b: 0.9, a: 0.05 } },
-    { offset: 1, color: { r: 0.75, g: 0.76, b: 0.8, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-  },
-};
-const ICE_BRICK_EMITTER_FILL2: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.85, g: 0.95, b: 1.0, a: 0.09 } },
-    { offset: 0.45, color: { r: 0.65, g: 0.85, b: 1.0, a: 0.05 } },
-    { offset: 1, color: { r: 0.45, g: 0.7, b: 1.0, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-  },
-};
-const COAL_BRICK_EMITTER_FILL2: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 0.3, g: 0.3, b: 0.32, a: 0.09 } },
-    { offset: 0.45, color: { r: 0.18, g: 0.18, b: 0.2, a: 0.05 } },
-    { offset: 1, color: { r: 0.1, g: 0.1, b: 0.12, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-  },
-};
-const MAGMA_BRICK_EMITTER_FILL2: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  stops: [
-    { offset: 0, color: { r: 1.0, g: 0.65, b: 0.2, a: 0.09 } },
-    { offset: 0.45, color: { r: 1.0, g: 0.45, b: 0.15, a: 0.05 } },
-    { offset: 1, color: { r: 0.9, g: 0.3, b: 0.1, a: 0.02 } },
-  ],
-  noise: {
-    colorAmplitude: 0.04,
-    alphaAmplitude: 0.09,
-    scale: 0.5,
-  },
-};
-
-const CRITICAL_HIT_EMITTER_FILL: SceneFill = {
-  fillType: FILL_TYPES.RADIAL_GRADIENT,
-  start: { x: 0, y: 0 },
-  end: 6,
-  stops: [
-    { offset: 0, color: { r: 1, g: 0.65, b: 0.65, a: 0.4 } },
-    { offset: 0.45, color: { r: 0.95, g: 0.28, b: 0.28, a: 0.3 } },
-    { offset: 1, color: { r: 0.7, g: 0.05, b: 0.05, a: 0 } },
-  ],
-};
-
-const DEFAULT_EMITTER: ParticleEmitterConfig = {
-  emissionDurationMs: 700,
-  particlesPerSecond: 260,
-  baseSpeed: 0.1,
-  speedVariation: 0.05,
-  particleLifetimeMs: 1_400,
-  fadeStartMs: 700,
-  sizeRange: { min: 1, max: 2 },
-  spawnRadius: { min: 0, max: 12 },
-  spawnRadiusMultiplier: 1.5,
-  color: { r: 1, g: 0.85, b: 0.55, a: 1 },
-  arc: Math.PI * 2,
-  direction: 0,
-  fill: DEFAULT_EMITTER_FILL,
-};
-
-const GRAY_BRICK_DAMAGE_EMITTER: ParticleEmitterConfig = {
-  emissionDurationMs: 140,
-  particlesPerSecond: 340,
-  baseSpeed: 0.04,
-  speedVariation: 0.01,
-  particleLifetimeMs: 650,
-  fadeStartMs: 280,
-  sizeRange: { min: 1, max: 2 },
-  spawnRadius: { min: 0, max: 5 },
-  spawnRadiusMultiplier: 1.2,
-  color: { r: 0.82, g: 0.84, b: 0.88, a: 1 },
-  arc: Math.PI * 2,
-  direction: 0,
-  fill: {
-    fillType: FILL_TYPES.SOLID,
-    color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
-  },
-  shape: "triangle",
-  sizeGrowthRate: 1.0,
-};
-
-const GRAY_BRICK_DESTRUCTION_EMITTER: ParticleEmitterConfig = {
-  emissionDurationMs: 220,
-  particlesPerSecond: 42,
-  baseSpeed: 0.05,
-  speedVariation: 0.01,
-  particleLifetimeMs: 750,
-  fadeStartMs: 480,
-  sizeRange: { min: 26, max: 49 },
-  spawnRadius: { min: 0, max: 8 },
-  spawnRadiusMultiplier: 1.5,
-  color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
-  arc: Math.PI * 2,
-  direction: 0,
-  fill: GRAY_BRICK_EMITTER_FILL,
-  shape: "circle",
-  sizeGrowthRate: 1.35,
-};
-
-
-const GRAY_BRICK_DESTRUCTION_EMITTER_V2: ParticleEmitterConfig = {
-  emissionDurationMs: 220,
-  particlesPerSecond: 720,
-  baseSpeed: 0.05,
-  speedVariation: 0.01,
-  particleLifetimeMs: 950,
-  fadeStartMs: 480,
-  sizeRange: { min: 1, max: 3 },
-  spawnRadius: { min: 0, max: 8 },
-  spawnRadiusMultiplier: 1.5,
-  color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
-  arc: Math.PI * 2,
-  direction: 0,
-  fill: {
-    fillType: FILL_TYPES.SOLID,
-    color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
-  },
-  shape: "triangle",
-  sizeGrowthRate: 1.0,
-  maxParticles: 1000,
-};
-
-const CRITICAL_HIT_EMITTER: ParticleEmitterConfig = {
-  emissionDurationMs: 240,
-  particlesPerSecond: 220,
-  baseSpeed: 0.16,
-  speedVariation: 0.05,
-  particleLifetimeMs: 520,
-  fadeStartMs: 220,
-  sizeRange: { min: 0.8, max: 1.6 },
-  spawnRadius: { min: 0, max: 6 },
-  spawnRadiusMultiplier: 1.1,
-  color: { r: 0.95, g: 0.2, b: 0.2, a: 1 },
-  arc: Math.PI * 2,
-  direction: 0,
-  fill: CRITICAL_HIT_EMITTER_FILL,
-};
-
-const EXPLOSION_DB: Record<ExplosionType, ExplosionConfig> = {
-  plasmoid: {
-    lifetimeMs: 3_000,
-    defaultInitialRadius: 12,
-    waves: createSimpleWave({
-      defaultInitialRadius: 12,
-      radiusExtension: 180,
-      startAlpha: 0.85,
-      endAlpha: 0,
-      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: DEFAULT_EMITTER,
-  },
-  smallCannon: {
-    lifetimeMs: 2_000,
-    defaultInitialRadius: 3,
-    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
-    waves: createSimpleWave({
-      defaultInitialRadius: 3,
-      radiusExtension: 10,
-      startAlpha: 0.45,
-      endAlpha: 0,
-      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      baseSpeed: 0.04,
-      speedVariation: 0.025,
-      fadeStartMs: 400,
-      particleLifetimeMs: 1_000,
-      particlesPerSecond: 2760,
-      sizeRange: { min: 0.5, max: 4.4 },
-      emissionDurationMs: 400,
-      spawnRadius: { min: 0, max: 0.1 },
-      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
-      // color: { r: 1, g: 1, b: 1, a: 1 },
-      fill: SMALL_CANNON_EMITTER_FILL,
-      radialVelocity: true, // Частинки рухаються від центру вибуху
-    },
-  },
-  smallCannonGrey: {
-    lifetimeMs: 2_000,
-    defaultInitialRadius: 3,
-    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
-    waves: createSimpleWave({
-      defaultInitialRadius: 3,
-      radiusExtension: 10,
-      startAlpha: 0.45,
-      endAlpha: 0,
-      gradientStops: GREY_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      baseSpeed: 0.04,
-      speedVariation: 0.025,
-      fadeStartMs: 400,
-      particleLifetimeMs: 1_000,
-      particlesPerSecond: 1760,
-      sizeRange: { min: 0.5, max: 3.4 },
-      emissionDurationMs: 400,
-      spawnRadius: { min: 0, max: 0.1 },
-      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
-      // color: { r: 1, g: 1, b: 1, a: 1 },
-      fill: SMALL_GREY_CANNON_EMITTER_FILL,
-      radialVelocity: true, // Частинки рухаються від центру вибуху
-    },
-  },
-  bigCannon: {
-    lifetimeMs: 2_000,
-    defaultInitialRadius: 3,
-    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
-    waves: createSimpleWave({
-      defaultInitialRadius: 3,
-      radiusExtension: 20,
-      startAlpha: 0.65,
-      endAlpha: 0,
-      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      baseSpeed: 0.06,
-      speedVariation: 0.035,
-      fadeStartMs: 600,
-      particleLifetimeMs: 1_200,
-      particlesPerSecond: 3760,
-      sizeRange: { min: 0.3, max: 4.4 },
-      emissionDurationMs: 400,
-      spawnRadius: { min: 0, max: 0.1 },
-      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
-      // color: { r: 1, g: 1, b: 1, a: 1 },
-      fill: SMALL_CANNON_EMITTER_FILL,
-      radialVelocity: true, // Частинки рухаються від центру вибуху
-    },
-  },
-  magnetic: {
-    lifetimeMs: 3_000,
-    defaultInitialRadius: 12,
-    waves: createSimpleWave({
-      defaultInitialRadius: 12,
-      radiusExtension: 180,
-      startAlpha: 0.45,
-      endAlpha: 0,
-      gradientStops: MAGNETIC_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      emissionDurationMs: 100,
-      particlesPerSecond: 300,
-      baseSpeed: 0.06,
-      speedVariation: 0.02,
-      particleLifetimeMs: 1_000,
-      sizeRange: { min: 25.6, max: 47.2 },
-      // color: { r: 1, g: 1, b: 1, a: 1 },
-      fill: MAGNETIC_EMITTER_FILL,
-    },
-  },
-  healWave: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 12,
-    waves: createSimpleWave({
-      defaultInitialRadius: 12,
-      radiusExtension: 120,
-      startAlpha: 0.85,
-      endAlpha: 0,
-      gradientStops: HEAL_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      emissionDurationMs: 0,
-      particlesPerSecond: 0,
-      baseSpeed: 0,
-      speedVariation: 0,
-      particleLifetimeMs: 0,
-      fadeStartMs: 0,
-      sizeRange: { min: 0, max: 0 },
-      spawnRadius: { min: 0, max: 0 },
-      spawnRadiusMultiplier: 1,
-      color: { r: 0.6, g: 1, b: 0.7, a: 1 },
-      arc: 0,
-      direction: 0,
-    },
-  },
-  fireball: {
-    lifetimeMs: 800,
-    defaultInitialRadius: 15,
-    waves: createSimpleWave({
-      defaultInitialRadius: 15,
-      radiusExtension: 50,
-      startAlpha: 0.9,
-      endAlpha: 0.3,
-      gradientStops: [
-        { offset: 0, color: { r: 1, g: 0.4, b: 0.1, a: 0.9 } },
-        { offset: 0.5, color: { r: 1, g: 0.7, b: 0.3, a: 0.6 } },
-        { offset: 1, color: { r: 1, g: 0.9, b: 0.5, a: 0 } },
-      ],
-    }),
-    emitter: {
-      emissionDurationMs: 500,
-      particlesPerSecond: 600,
-      baseSpeed: 0.08,
-      speedVariation: 0.02,
-      particleLifetimeMs: 1100,
-      fadeStartMs: 600,
-      sizeRange: { min: 16, max: 46 },
-      spawnRadius: { min: 0, max: 5 },
-      spawnRadiusMultiplier: 1,
-      sizeGrowthRate: 3,
-      color: { r: 1, g: 0.6, b: 0.2, a: 1 },
-      arc: Math.PI * 2,
-      direction: 0,
-      fill: {
-        fillType: FILL_TYPES.RADIAL_GRADIENT,
-        start: { x: 0, y: 0 },
-        stops: [
-          { offset: 0, color: { r: 1, g: 0.8, b: 0.6, a: 0.03 } },
-          { offset: 0.5, color: { r: 1, g: 0.7, b: 0.4, a: 0.05 } },
-          { offset: 1, color: { r: 1, g: 0.7, b: 0.3, a: 0 } },
-        ],
-        noise: {
-          colorAmplitude: 0.01,
-          alphaAmplitude: 0.03,
-          scale: 0.5,
-        },
-      },
-      radialVelocity: true,
-    },
-  },
-  grayBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 40,
-      startAlpha: 0.4,
-      endAlpha: 0.0,
-      gradientStops: GRAY_BRICK_HIT_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: GRAY_BRICK_DAMAGE_EMITTER,
-  },
-  grayBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 60,
-      startAlpha: 0.6,
-      endAlpha: 0.0,
-      gradientStops: GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: GRAY_BRICK_DESTRUCTION_EMITTER_V2,
-  },
-  grayBrickDestroyV2: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 60,
-      startAlpha: 0.5,
-      endAlpha: 0.0,
-      gradientStops: GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: GRAY_BRICK_DESTRUCTION_EMITTER_V2,
-  },
-  yellowBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 40,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: YELLOW_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 1, g: 0.95, b: 0.6, a: 1 }, fill: YELLOW_BRICK_EMITTER_FILL },
-  },
-  yellowBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 60,
-      startAlpha: 0.7,
-      endAlpha: 0,
-      gradientStops: YELLOW_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 1, g: 0.95, b: 0.6, a: 1 }, fill: YELLOW_BRICK_EMITTER_FILL },
-  },
-  organicBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: GREEN_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.6, g: 1, b: 0.7, a: 1 }, fill: GREEN_BRICK_EMITTER_FILL },
-  },
-  organicBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.7,
-      endAlpha: 0,
-      gradientStops: GREEN_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.6, g: 1, b: 0.7, a: 1 }, fill: GREEN_BRICK_EMITTER_FILL },
-  },
-  ironBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: ORANGE_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 1, g: 0.8, b: 0.55, a: 1 }, fill: ORANGE_BRICK_EMITTER_FILL },
-  },
-  ironBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.7,
-      endAlpha: 0,
-      gradientStops: ORANGE_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 1, g: 0.8, b: 0.55, a: 1 }, fill: ORANGE_BRICK_EMITTER_FILL },
-  },
-  woodBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: BROWN_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.75, g: 0.55, b: 0.35, a: 1 }, fill: BROWN_BRICK_EMITTER_FILL },
-  },
-  woodBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.8,
-      endAlpha: 0,
-      gradientStops: BROWN_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.75, g: 0.55, b: 0.35, a: 1 }, fill: BROWN_BRICK_EMITTER_FILL },
-  },
-  copperBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.5,
-      endAlpha: 0,
-      gradientStops: ORANGE_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.95, g: 0.65, b: 0.35, a: 1 }, fill: ORANGE_BRICK_EMITTER_FILL },
-  },
-  copperBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.7,
-      endAlpha: 0,
-      gradientStops: ORANGE_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.95, g: 0.65, b: 0.35, a: 1 }, fill: ORANGE_BRICK_EMITTER_FILL },
-  },
-  silverBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: SILVER_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.92, g: 0.93, b: 0.96, a: 1 }, fill: SILVER_BRICK_EMITTER_FILL2 },
-  },
-  silverBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.8,
-      endAlpha: 0,
-      gradientStops: SILVER_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.92, g: 0.93, b: 0.96, a: 1 }, fill: SILVER_BRICK_EMITTER_FILL2 },
-  },
-  coalBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: COAL_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.24, g: 0.24, b: 0.27, a: 1 }, fill: COAL_BRICK_EMITTER_FILL2 },
-  },
-  coalBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.8,
-      endAlpha: 0,
-      gradientStops: COAL_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.24, g: 0.24, b: 0.27, a: 1 }, fill: COAL_BRICK_EMITTER_FILL2 },
-  },
-  iceBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: ICE_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 0.7, g: 0.9, b: 1.0, a: 1 }, fill: ICE_BRICK_EMITTER_FILL2 },
-  },
-  iceBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.8,
-      endAlpha: 0,
-      gradientStops: ICE_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 0.85, g: 0.95, b: 1.0, a: 1 }, fill: ICE_BRICK_EMITTER_FILL2 },
-  },
-  magmaBrickHit: {
-    lifetimeMs: 1_000,
-    defaultInitialRadius: 6,
-    waves: createSimpleWave({
-      defaultInitialRadius: 6,
-      radiusExtension: 30,
-      startAlpha: 0.6,
-      endAlpha: 0,
-      gradientStops: MAGMA_BRICK_HIT_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DAMAGE_EMITTER, color: { r: 1.0, g: 0.5, b: 0.15, a: 1 }, fill: MAGMA_BRICK_EMITTER_FILL2 },
-  },
-  magmaBrickDestroy: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 50,
-      startAlpha: 0.8,
-      endAlpha: 0,
-      gradientStops: MAGMA_BRICK_DESTROY_WAVE,
-    }),
-    emitter: { ...GRAY_BRICK_DESTRUCTION_EMITTER, color: { r: 1.0, g: 0.65, b: 0.2, a: 1 }, fill: MAGMA_BRICK_EMITTER_FILL2 },
-  },
-  criticalHit: {
-    lifetimeMs: 900,
-    defaultInitialRadius: 10,
-    waves: createSimpleWave({
-      defaultInitialRadius: 10,
-      radiusExtension: 70,
-      startAlpha: 0.9,
-      endAlpha: 0,
-      gradientStops: CRITICAL_HIT_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: CRITICAL_HIT_EMITTER,
-  },
-  weakenCurse: {
-    lifetimeMs: 3_600,
-    defaultInitialRadius: 10,
-    waves: [
-      {
-        initialInnerRadius: 0,
-        expansionInnerRadius: 10,
-        initialOuterRadius: 10,
-        expansionOuterRadius: 50,
-        startAlpha: 0.9,
-        endAlpha: 0.1,
-        gradientStops: [
-          { offset: 0, color: { r: 0.4, g: 0.1, b: 0.5, a: 0.2 } },
-          { offset: 0.35, color: { r: 0.45, g: 0.15, b: 0.55, a: 0.5 } },
-          { offset: 0.75, color: { r: 0.5, g: 0.2, b: 0.6, a: 0.7 } },
-          { offset: 1, color: { r: 0.45, g: 0.25, b: 0.5, a: 0.0 } },
-        ] as const,
-      },
-      {
-        initialInnerRadius: 0,
-        expansionInnerRadius: 60,
-        initialOuterRadius: 50,
-        expansionOuterRadius: 110,
-        startAlpha: 0.9,
-        endAlpha: 0.1,
-        gradientStops: [
-          { offset: 0, color: { r: 0.4, g: 0.1, b: 0.5, a: 0.2 } },
-          { offset: 0.35, color: { r: 0.45, g: 0.15, b: 0.55, a: 0.5 } },
-          { offset: 0.75, color: { r: 0.5, g: 0.2, b: 0.6, a: 0.7 } },
-          { offset: 1, color: { r: 0.45, g: 0.25, b: 0.5, a: 0.0 } },
-        ] as const,
-        filaments: {
-          colorContrast: 0.1,
-          alphaContrast: 0.08,
-          width: 0.35,
-          density: 6.1,
-          edgeBlur: 0.35,
-        },
-      },
-    ],
-    emitter: {
-      emissionDurationMs: 0.220,
-      particlesPerSecond: 62,
-      baseSpeed: 0.05,
-      speedVariation: 0.01,
-      particleLifetimeMs: 750,
-      fadeStartMs: 480,
-      sizeRange: { min: 36, max: 59 },
-      spawnRadius: { min: 0, max: 8 },
-      spawnRadiusMultiplier: 1.5,
-      color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
-      arc: Math.PI * 2,
-      direction: 0,
-      fill: GRAY_BRICK_EMITTER_FILL,
-      shape: "circle",
-      sizeGrowthRate: 2.35,
-    },
-  },
-  smallLaser: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 3,
-    waves: createSimpleWave({
-      defaultInitialRadius: 2,
-      radiusExtension: 7,
-      startAlpha: 0.65,
-      endAlpha: 0,
-      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      baseSpeed: 0.04,
-      speedVariation: 0.025,
-      fadeStartMs: 400,
-      particleLifetimeMs: 1_000,
-      particlesPerSecond: 1760,
-      sizeRange: { min: 0.25, max: 1.4 },
-      emissionDurationMs: 400,
-      spawnRadius: { min: 0, max: 0.1 },
-      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
-      // color: { r: 1, g: 1, b: 1, a: 1 },
-      fill: SMALL_LASER_EMITTER_FILL,
-      radialVelocity: true, // Частинки рухаються від центру вибуху
-    },
-  },
-  smallEnergetic: {
-    lifetimeMs: 1_200,
-    defaultInitialRadius: 3,
-    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
-    waves: createSimpleWave({
-      defaultInitialRadius: 2,
-      radiusExtension: 5,
-      startAlpha: 0.95,
-      endAlpha: 0,
-      gradientStops: SMALL_ENERGETIC_WAVE_GRADIENT_STOPS,
-    }),
-    emitter: {
-      ...DEFAULT_EMITTER,
-      baseSpeed: 0.2,
-      speedVariation: 0.05,
-      fadeStartMs: 600,
-      particleLifetimeMs: 1_000,
-      particlesPerSecond: 1760,
-      sizeRange: { min: 0.25, max: 1.4 },
-      emissionDurationMs: 400,
-      spawnRadius: { min: 0, max: 0.2 },
-      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
-      fill: SMALL_ENERGETIC_EMITTER_FILL,
-      radialVelocity: true, // Частинки рухаються від центру вибуху
-    },
-  },
-};
-
-export const getExplosionConfig = (type: ExplosionType): ExplosionConfig => {
-  const config = EXPLOSION_DB[type];
-  if (!config) {
-    throw new Error(`Unknown explosion type: ${type}`);
-  }
-  return config;
-};
-
-export const EXPLOSION_TYPES = Object.keys(EXPLOSION_DB) as ExplosionType[];
+export * from "./explosions";
