+++ src/db/explosions/explosions.bricks.ts
@@ -1,0 +1,362 @@
+import type { ExplosionConfig, ExplosionType } from "./explosions.types";
+import { createSimpleWave } from "./explosions.helpers";
+import {
+  BROWN_BRICK_DESTROY_WAVE,
+  BROWN_BRICK_HIT_WAVE,
+  COAL_BRICK_DESTROY_WAVE,
+  COAL_BRICK_HIT_WAVE,
+  GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS,
+  GRAY_BRICK_HIT_WAVE_GRADIENT_STOPS,
+  GREEN_BRICK_DESTROY_WAVE,
+  GREEN_BRICK_HIT_WAVE,
+  ICE_BRICK_DESTROY_WAVE,
+  ICE_BRICK_HIT_WAVE,
+  MAGMA_BRICK_DESTROY_WAVE,
+  MAGMA_BRICK_HIT_WAVE,
+  ORANGE_BRICK_DESTROY_WAVE,
+  ORANGE_BRICK_HIT_WAVE,
+  SILVER_BRICK_DESTROY_WAVE,
+  SILVER_BRICK_HIT_WAVE,
+  YELLOW_BRICK_DESTROY_WAVE,
+  YELLOW_BRICK_HIT_WAVE,
+} from "./explosions.colors.const";
+import {
+  BROWN_BRICK_EMITTER_FILL,
+  COAL_BRICK_EMITTER_FILL2,
+  GRAY_BRICK_DAMAGE_EMITTER,
+  GRAY_BRICK_DESTRUCTION_EMITTER,
+  GRAY_BRICK_DESTRUCTION_EMITTER_V2,
+  GREEN_BRICK_EMITTER_FILL,
+  ICE_BRICK_EMITTER_FILL2,
+  MAGMA_BRICK_EMITTER_FILL2,
+  ORANGE_BRICK_EMITTER_FILL,
+  SILVER_BRICK_EMITTER_FILL2,
+  YELLOW_BRICK_EMITTER_FILL,
+} from "./explosions.emitters.const";
+
+export const BRICK_EXPLOSIONS: Partial<Record<ExplosionType, ExplosionConfig>> = {
+  grayBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 40,
+      startAlpha: 0.4,
+      endAlpha: 0.0,
+      gradientStops: GRAY_BRICK_HIT_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: GRAY_BRICK_DAMAGE_EMITTER,
+  },
+  grayBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 60,
+      startAlpha: 0.6,
+      endAlpha: 0.0,
+      gradientStops: GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: GRAY_BRICK_DESTRUCTION_EMITTER_V2,
+  },
+  grayBrickDestroyV2: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 60,
+      startAlpha: 0.5,
+      endAlpha: 0.0,
+      gradientStops: GRAY_BRICK_DESTROY_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: GRAY_BRICK_DESTRUCTION_EMITTER_V2,
+  },
+  yellowBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 40,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: YELLOW_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 1, g: 0.95, b: 0.6, a: 1 },
+      fill: YELLOW_BRICK_EMITTER_FILL,
+    },
+  },
+  yellowBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 60,
+      startAlpha: 0.7,
+      endAlpha: 0,
+      gradientStops: YELLOW_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 1, g: 0.95, b: 0.6, a: 1 },
+      fill: YELLOW_BRICK_EMITTER_FILL,
+    },
+  },
+  organicBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: GREEN_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.6, g: 1, b: 0.7, a: 1 },
+      fill: GREEN_BRICK_EMITTER_FILL,
+    },
+  },
+  organicBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.7,
+      endAlpha: 0,
+      gradientStops: GREEN_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.6, g: 1, b: 0.7, a: 1 },
+      fill: GREEN_BRICK_EMITTER_FILL,
+    },
+  },
+  ironBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: ORANGE_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 1, g: 0.8, b: 0.55, a: 1 },
+      fill: ORANGE_BRICK_EMITTER_FILL,
+    },
+  },
+  ironBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.7,
+      endAlpha: 0,
+      gradientStops: ORANGE_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 1, g: 0.8, b: 0.55, a: 1 },
+      fill: ORANGE_BRICK_EMITTER_FILL,
+    },
+  },
+  woodBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: BROWN_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.75, g: 0.55, b: 0.35, a: 1 },
+      fill: BROWN_BRICK_EMITTER_FILL,
+    },
+  },
+  woodBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.8,
+      endAlpha: 0,
+      gradientStops: BROWN_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.75, g: 0.55, b: 0.35, a: 1 },
+      fill: BROWN_BRICK_EMITTER_FILL,
+    },
+  },
+  copperBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.5,
+      endAlpha: 0,
+      gradientStops: ORANGE_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.95, g: 0.65, b: 0.35, a: 1 },
+      fill: ORANGE_BRICK_EMITTER_FILL,
+    },
+  },
+  copperBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.7,
+      endAlpha: 0,
+      gradientStops: ORANGE_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.95, g: 0.65, b: 0.35, a: 1 },
+      fill: ORANGE_BRICK_EMITTER_FILL,
+    },
+  },
+  silverBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: SILVER_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.92, g: 0.93, b: 0.96, a: 1 },
+      fill: SILVER_BRICK_EMITTER_FILL2,
+    },
+  },
+  silverBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.8,
+      endAlpha: 0,
+      gradientStops: SILVER_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.92, g: 0.93, b: 0.96, a: 1 },
+      fill: SILVER_BRICK_EMITTER_FILL2,
+    },
+  },
+  coalBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: COAL_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.24, g: 0.24, b: 0.27, a: 1 },
+      fill: COAL_BRICK_EMITTER_FILL2,
+    },
+  },
+  coalBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.8,
+      endAlpha: 0,
+      gradientStops: COAL_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.24, g: 0.24, b: 0.27, a: 1 },
+      fill: COAL_BRICK_EMITTER_FILL2,
+    },
+  },
+  iceBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: ICE_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 0.7, g: 0.9, b: 1.0, a: 1 },
+      fill: ICE_BRICK_EMITTER_FILL2,
+    },
+  },
+  iceBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.8,
+      endAlpha: 0,
+      gradientStops: ICE_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 0.85, g: 0.95, b: 1.0, a: 1 },
+      fill: ICE_BRICK_EMITTER_FILL2,
+    },
+  },
+  magmaBrickHit: {
+    lifetimeMs: 1_000,
+    defaultInitialRadius: 6,
+    waves: createSimpleWave({
+      defaultInitialRadius: 6,
+      radiusExtension: 30,
+      startAlpha: 0.6,
+      endAlpha: 0,
+      gradientStops: MAGMA_BRICK_HIT_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DAMAGE_EMITTER,
+      color: { r: 1.0, g: 0.5, b: 0.15, a: 1 },
+      fill: MAGMA_BRICK_EMITTER_FILL2,
+    },
+  },
+  magmaBrickDestroy: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 50,
+      startAlpha: 0.8,
+      endAlpha: 0,
+      gradientStops: MAGMA_BRICK_DESTROY_WAVE,
+    }),
+    emitter: {
+      ...GRAY_BRICK_DESTRUCTION_EMITTER,
+      color: { r: 1.0, g: 0.65, b: 0.2, a: 1 },
+      fill: MAGMA_BRICK_EMITTER_FILL2,
+    },
+  },
+};
