+++ src/db/explosions/explosions.projectiles.ts
@@ -1,0 +1,313 @@
+import { FILL_TYPES } from "@core/logic/provided/services/scene-object-manager/scene-object-manager.const";
+import { getAssetUrl } from "@shared/helpers/assets.helper";
+import type { ExplosionConfig, ExplosionType } from "./explosions.types";
+import { createSimpleWave } from "./explosions.helpers";
+import {
+  CRITICAL_HIT_WAVE_GRADIENT_STOPS,
+  GREY_WAVE_GRADIENT_STOPS,
+  HEAL_WAVE_GRADIENT_STOPS,
+  MAGNETIC_WAVE_GRADIENT_STOPS,
+  PLASMOID_WAVE_GRADIENT_STOPS,
+  SMALL_ENERGETIC_WAVE_GRADIENT_STOPS,
+  WEAKEN_CURSE_WAVE_GRADIENT_STOPS,
+} from "./explosions.colors.const";
+import {
+  CRITICAL_HIT_EMITTER,
+  DEFAULT_EMITTER,
+  GRAY_BRICK_EMITTER_FILL,
+  MAGNETIC_EMITTER_FILL,
+  SMALL_CANNON_EMITTER_FILL,
+  SMALL_ENERGETIC_EMITTER_FILL,
+  SMALL_GREY_CANNON_EMITTER_FILL,
+  SMALL_LASER_EMITTER_FILL,
+  WEAKEN_CURSE_FILAMENTS,
+} from "./explosions.emitters.const";
+
+export const PROJECTILE_EXPLOSIONS: Partial<Record<ExplosionType, ExplosionConfig>> = {
+  plasmoid: {
+    lifetimeMs: 3_000,
+    defaultInitialRadius: 12,
+    waves: createSimpleWave({
+      defaultInitialRadius: 12,
+      radiusExtension: 180,
+      startAlpha: 0.85,
+      endAlpha: 0,
+      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: DEFAULT_EMITTER,
+  },
+  smallCannon: {
+    lifetimeMs: 2_000,
+    defaultInitialRadius: 3,
+    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
+    waves: createSimpleWave({
+      defaultInitialRadius: 3,
+      radiusExtension: 10,
+      startAlpha: 0.45,
+      endAlpha: 0,
+      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      baseSpeed: 0.04,
+      speedVariation: 0.025,
+      fadeStartMs: 400,
+      particleLifetimeMs: 1_000,
+      particlesPerSecond: 2760,
+      sizeRange: { min: 0.5, max: 4.4 },
+      emissionDurationMs: 400,
+      spawnRadius: { min: 0, max: 0.1 },
+      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
+      fill: SMALL_CANNON_EMITTER_FILL,
+      radialVelocity: true, // Частинки рухаються від центру вибуху
+    },
+  },
+  smallCannonGrey: {
+    lifetimeMs: 2_000,
+    defaultInitialRadius: 3,
+    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
+    waves: createSimpleWave({
+      defaultInitialRadius: 3,
+      radiusExtension: 10,
+      startAlpha: 0.45,
+      endAlpha: 0,
+      gradientStops: GREY_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      baseSpeed: 0.04,
+      speedVariation: 0.025,
+      fadeStartMs: 400,
+      particleLifetimeMs: 1_000,
+      particlesPerSecond: 1760,
+      sizeRange: { min: 0.5, max: 3.4 },
+      emissionDurationMs: 400,
+      spawnRadius: { min: 0, max: 0.1 },
+      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
+      fill: SMALL_GREY_CANNON_EMITTER_FILL,
+      radialVelocity: true, // Частинки рухаються від центру вибуху
+    },
+  },
+  bigCannon: {
+    lifetimeMs: 2_000,
+    defaultInitialRadius: 3,
+    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
+    waves: createSimpleWave({
+      defaultInitialRadius: 3,
+      radiusExtension: 20,
+      startAlpha: 0.65,
+      endAlpha: 0,
+      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      baseSpeed: 0.06,
+      speedVariation: 0.035,
+      fadeStartMs: 600,
+      particleLifetimeMs: 1_200,
+      particlesPerSecond: 3760,
+      sizeRange: { min: 0.3, max: 4.4 },
+      emissionDurationMs: 400,
+      spawnRadius: { min: 0, max: 0.1 },
+      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
+      fill: SMALL_CANNON_EMITTER_FILL,
+      radialVelocity: true, // Частинки рухаються від центру вибуху
+    },
+  },
+  magnetic: {
+    lifetimeMs: 3_000,
+    defaultInitialRadius: 12,
+    waves: createSimpleWave({
+      defaultInitialRadius: 12,
+      radiusExtension: 180,
+      startAlpha: 0.45,
+      endAlpha: 0,
+      gradientStops: MAGNETIC_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      emissionDurationMs: 100,
+      particlesPerSecond: 300,
+      baseSpeed: 0.06,
+      speedVariation: 0.02,
+      particleLifetimeMs: 1_000,
+      sizeRange: { min: 25.6, max: 47.2 },
+      fill: MAGNETIC_EMITTER_FILL,
+    },
+  },
+  healWave: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 12,
+    waves: createSimpleWave({
+      defaultInitialRadius: 12,
+      radiusExtension: 120,
+      startAlpha: 0.85,
+      endAlpha: 0,
+      gradientStops: HEAL_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      emissionDurationMs: 0,
+      particlesPerSecond: 0,
+      baseSpeed: 0,
+      speedVariation: 0,
+      particleLifetimeMs: 0,
+      fadeStartMs: 0,
+      sizeRange: { min: 0, max: 0 },
+      spawnRadius: { min: 0, max: 0 },
+      spawnRadiusMultiplier: 1,
+      color: { r: 0.6, g: 1, b: 0.7, a: 1 },
+      arc: 0,
+      direction: 0,
+    },
+  },
+  fireball: {
+    lifetimeMs: 800,
+    defaultInitialRadius: 15,
+    waves: createSimpleWave({
+      defaultInitialRadius: 15,
+      radiusExtension: 50,
+      startAlpha: 0.9,
+      endAlpha: 0.3,
+      gradientStops: [
+        { offset: 0, color: { r: 1, g: 0.4, b: 0.1, a: 0.9 } },
+        { offset: 0.5, color: { r: 1, g: 0.7, b: 0.3, a: 0.6 } },
+        { offset: 1, color: { r: 1, g: 0.9, b: 0.5, a: 0 } },
+      ],
+    }),
+    emitter: {
+      emissionDurationMs: 500,
+      particlesPerSecond: 600,
+      baseSpeed: 0.08,
+      speedVariation: 0.02,
+      particleLifetimeMs: 1100,
+      fadeStartMs: 600,
+      sizeRange: { min: 16, max: 46 },
+      spawnRadius: { min: 0, max: 5 },
+      spawnRadiusMultiplier: 1,
+      sizeGrowthRate: 3,
+      color: { r: 1, g: 0.6, b: 0.2, a: 1 },
+      arc: Math.PI * 2,
+      direction: 0,
+      fill: {
+        fillType: FILL_TYPES.RADIAL_GRADIENT,
+        start: { x: 0, y: 0 },
+        stops: [
+          { offset: 0, color: { r: 1, g: 0.8, b: 0.6, a: 0.03 } },
+          { offset: 0.5, color: { r: 1, g: 0.7, b: 0.4, a: 0.05 } },
+          { offset: 1, color: { r: 1, g: 0.7, b: 0.3, a: 0 } },
+        ],
+        noise: {
+          colorAmplitude: 0.01,
+          alphaAmplitude: 0.03,
+          scale: 0.5,
+        },
+      },
+      radialVelocity: true,
+    },
+  },
+  criticalHit: {
+    lifetimeMs: 900,
+    defaultInitialRadius: 10,
+    waves: createSimpleWave({
+      defaultInitialRadius: 10,
+      radiusExtension: 70,
+      startAlpha: 0.9,
+      endAlpha: 0,
+      gradientStops: CRITICAL_HIT_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: CRITICAL_HIT_EMITTER,
+  },
+  weakenCurse: {
+    lifetimeMs: 3_600,
+    defaultInitialRadius: 10,
+    waves: [
+      {
+        initialInnerRadius: 0,
+        expansionInnerRadius: 10,
+        initialOuterRadius: 10,
+        expansionOuterRadius: 50,
+        startAlpha: 0.9,
+        endAlpha: 0.1,
+        gradientStops: WEAKEN_CURSE_WAVE_GRADIENT_STOPS,
+      },
+      {
+        initialInnerRadius: 0,
+        expansionInnerRadius: 60,
+        initialOuterRadius: 50,
+        expansionOuterRadius: 110,
+        startAlpha: 0.9,
+        endAlpha: 0.1,
+        gradientStops: WEAKEN_CURSE_WAVE_GRADIENT_STOPS,
+        filaments: WEAKEN_CURSE_FILAMENTS,
+      },
+    ],
+    emitter: {
+      emissionDurationMs: 0.220,
+      particlesPerSecond: 62,
+      baseSpeed: 0.05,
+      speedVariation: 0.01,
+      particleLifetimeMs: 750,
+      fadeStartMs: 480,
+      sizeRange: { min: 36, max: 59 },
+      spawnRadius: { min: 0, max: 8 },
+      spawnRadiusMultiplier: 1.5,
+      color: { r: 0.85, g: 0.87, b: 0.92, a: 1 },
+      arc: Math.PI * 2,
+      direction: 0,
+      fill: GRAY_BRICK_EMITTER_FILL,
+      shape: "circle",
+      sizeGrowthRate: 2.35,
+    },
+  },
+  smallLaser: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 3,
+    waves: createSimpleWave({
+      defaultInitialRadius: 2,
+      radiusExtension: 7,
+      startAlpha: 0.65,
+      endAlpha: 0,
+      gradientStops: PLASMOID_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      baseSpeed: 0.04,
+      speedVariation: 0.025,
+      fadeStartMs: 400,
+      particleLifetimeMs: 1_000,
+      particlesPerSecond: 1760,
+      sizeRange: { min: 0.25, max: 1.4 },
+      emissionDurationMs: 400,
+      spawnRadius: { min: 0, max: 0.1 },
+      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
+      fill: SMALL_LASER_EMITTER_FILL,
+      radialVelocity: true, // Частинки рухаються від центру вибуху
+    },
+  },
+  smallEnergetic: {
+    lifetimeMs: 1_200,
+    defaultInitialRadius: 3,
+    soundEffectUrl: getAssetUrl("audio/sounds/general_effects/explosion.mp3"),
+    waves: createSimpleWave({
+      defaultInitialRadius: 2,
+      radiusExtension: 5,
+      startAlpha: 0.95,
+      endAlpha: 0,
+      gradientStops: SMALL_ENERGETIC_WAVE_GRADIENT_STOPS,
+    }),
+    emitter: {
+      ...DEFAULT_EMITTER,
+      baseSpeed: 0.2,
+      speedVariation: 0.05,
+      fadeStartMs: 600,
+      particleLifetimeMs: 1_000,
+      particlesPerSecond: 1760,
+      sizeRange: { min: 0.25, max: 1.4 },
+      emissionDurationMs: 400,
+      spawnRadius: { min: 0, max: 0.2 },
+      spawnRadiusMultiplier: undefined, // Override DEFAULT_EMITTER to use explicit spawnRadius
+      fill: SMALL_ENERGETIC_EMITTER_FILL,
+      radialVelocity: true, // Частинки рухаються від центру вибуху
+    },
+  },
+};
